+++ example/lib/main_wav_stream.dart
@@ -70,7 +70,7 @@ class _MyHomePageState extends State<MyHomePage> {
 
   Future<void> start() async {
     try {
-      await SoLoud.instance.initialize();
+      await SoLoud.instance.init();
     } catch (e) {
       debugPrint('isolate starting error: $e');
       return;
@@ -110,7 +110,7 @@ class _MyHomePageState extends State<MyHomePage> {
   }
 
   Future<void> stop() async {
-    await SoLoud.instance.shutdown();
+    SoLoud.instance.deinit();
     SoLoudCapture.instance.stopCapture();
     sounds.clear();
   }